Split out the WebSocket client
This allows the client to be used without
pulling in vertx-http.
