+++ extensions/websockets/client/runtime/src/main/java/io/quarkus/undertow/websockets/client/runtime/ExecutorSupplier.java
@@ -1,0 +1,14 @@
+package io.quarkus.undertow.websockets.client.runtime;
+
+import java.util.concurrent.Executor;
+import java.util.function.Supplier;
+
+public class ExecutorSupplier implements Supplier<Executor> {
+
+    static volatile Executor executor;
+
+    @Override
+    public Executor get() {
+        return executor;
+    }
+}
