+++ extensions/websockets/client/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,15 @@
+name: "WebSockets Client"
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+metadata:
+  short-name: "websockets-client"
+  keywords:
+  - "websocket"
+  - "websockets"
+  - "web-socket"
+  - "web-sockets"
+  - "http"
+  - "client"
+  guide: "https://quarkus.io/guides/websockets"
+  categories:
+  - "web"
+  status: "stable"
