--- extensions/websockets/runtime/src/main/java/io/quarkus/undertow/websockets/runtime/ExecutorSupplier.java
@@ -1,14 +1,0 @@
-package io.quarkus.undertow.websockets.runtime;
-
-import java.util.concurrent.Executor;
-import java.util.function.Supplier;
-
-public class ExecutorSupplier implements Supplier<Executor> {
-
-    static volatile Executor executor;
-
-    @Override
-    public Executor get() {
-        return executor;
-    }
-}
