+++ extensions/websockets/server/pom.xml
@@ -6,17 +6,16 @@
         <artifactId>quarkus-extensions-parent</artifactId>
         <groupId>io.quarkus</groupId>
         <version>999-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
+        <relativePath>../../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>quarkus-websockets-aggregator</artifactId>
-    <name>Quarkus - WebSockets Aggregator</name>
+    <artifactId>quarkus-websockets-parent</artifactId>
+    <name>Quarkus - WebSockets</name>
     <packaging>pom</packaging>
     <modules>
-        <module>server</module>
-        <module>client</module>
-        <module>legacy</module>
+        <module>deployment</module>
+        <module>runtime</module>
     </modules>
 
 </project>
