+++ extensions/websockets/server/runtime/src/main/resources/META-INF/quarkus-extension.yaml
